backend/requirements.txt
# === Core ===
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.*
pyyaml==6.*
python-dotenv==1.*

# === Agent Runtime ===
langgraph>=0.2.60
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# === MCP (动态工具发现) ===
langchain-mcp-adapters==0.1.*

# === LLM Gateway ===
litellm==1.*

# === Database ===
asyncpg==0.30.*
sqlalchemy[asyncio]==2.*
psycopg[binary]==3.*
langgraph-checkpoint-postgres>=2.0.0
alembic==1.*
pgvector==0.3.*

# === Redis ===
redis[hiredis]==5.*

# === Scheduler ===
apscheduler==3.*

# === Auth ===
PyJWT==2.*
bcrypt==4.*
python-multipart==0.0.*

# === Push Notifications ===
pywebpush==2.*

# === Utilities ===
httpx==0.28.*
structlog==24.*
slowapi==0.1.*

# === Testing ===
pytest==8.*
pytest-asyncio==0.24.*
pytest-mock==3.*
pytest-cov==5.*
aiosqlite==0.20.*
